chore: GC pt3: further wiping

+++ depchain-common/src/main/java/tecnico/depchain/depchain_common/blockchain/Transaction.java
@@ -4,14 +4,13 @@ import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.ObjectOutputStream;
 import java.io.Serializable;
-import java.math.BigInteger;
 
 import org.apache.tuweni.bytes.Bytes;
 import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.datatypes.Wei;
 
 public record Transaction(
-		BigInteger nonce,
+		long nonce,
 		Address from,
 		Address to,
 		Wei gasPrice,
